Login screens: mutualize the code which create session from the credentials
